requirements.txt
# Function dependencies, for example:
# package>=version
numpy
Pillow
tensorflow
google-cloud-storage
